# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: HaulBrokr API
  version: 1.0.0
  description: HaulBrokr API — Uber for dump trucks
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /requests/{requestId}/bids:
    get:
      operationId: listBids
      tags: [bids]
      summary: List bids on a job request
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of bids
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bid"
    post:
      operationId: createBid
      tags: [bids]
      summary: Place a bid on a job request
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BidInput"
      responses:
        "201":
          description: Bid placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bid"
  /bids/{id}:
    get:
      operationId: getBid
      tags: [bids]
      summary: Get a bid by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bid found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bid"
        "404":
          description: Bid not found
    patch:
      operationId: updateBid
      tags: [bids]
      summary: Accept or reject a bid
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BidUpdate"
      responses:
        "200":
          description: Bid updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bid"

  # ── Jobs ─────────────────────────────────────────────────────────────────────
...
  /factoring:
    get:
      operationId: listFactoringRequests
      tags: [factoring]
      summary: List provider factoring requests
      responses:
        "200":
          description: Factoring requests
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
    post:
      operationId: createFactoringRequest
      tags: [factoring]
      summary: Request invoice factoring advance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [jobId]
              properties:
                jobId:
                  type: integer
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
  /factoring/{id}/approve:
    patch:
      operationId: approveFactoringRequest
      tags: [factoring]
      summary: Staff approve and fund factoring request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                type: object

  # ── QuickBooks (simulated) ─────────────────────────────────────────────────
...
  /jobs/{jobId}/evidence:
    get:
      operationId: listJobEvidence
      tags: [evidence]
      summary: List delivery evidence for a job
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Evidence list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
    post:
      operationId: createJobEvidence
      tags: [evidence]
      summary: Upload delivery evidence
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object

  # ── Bin orders (customer) ──────────────────────────────────────────────────
...
components:
  schemas:
...
    Bid:
      type: object
      required: [id, requestId, providerId, providerCompany, ratePerHour, trucksOffered, status, createdAt]
      properties:
        id:
          type: integer
        requestId:
          type: integer
        providerId:
          type: integer
        providerCompany:
          type: string
        ratePerHour:
          type: number
        trucksOffered:
          type: integer
        estimatedHours:
          type: ["number", "null"]
        message:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, accepted, rejected, withdrawn]
        createdAt:
          type: string
          format: date-time
    BidInput:
      type: object
      required: [ratePerHour, trucksOffered]
      properties:
        ratePerHour:
          type: number
        trucksOffered:
          type: integer
        estimatedHours:
          type: number
        message:
          type: string
    BidUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [accepted, rejected, withdrawn]

    # ── Job ───────────────────────────────────────────────────────────────────
...